Refactor Snippet and Progress components. Added folder "ui/", here will be, components for blogging
