+++ app/components/Progress/cva.ts
@@ -3,9 +3,9 @@ import { cva } from 'class-variance-authority'
 export const cvaProgress = cva('progress', {
   variants: {
     status: {
-      success: ['h-4 rounded bg-success-300'],
-      warning: ['h-4 rounded bg-warning-300'],
-      error: ['h-4 rounded bg-error-300']
+      success: ['h-4 rounded-[0.2rem] bg-success-300'],
+      warning: ['h-4 rounded-[0.2rem] bg-warning-300'],
+      error: ['h-4 rounded-[0.2rem] bg-error-300']
     }
   },
   defaultVariants: {
